src/backend/requirements.txt
cython
hypercorn[uvloop]
aiomysql
pyyaml
